+++ extras/integration_tests/tests/lib/suite/custom_objects.dart
@@ -1,3 +1,4 @@
+import 'package:moor/moor.dart' hide isNull;
 import 'package:test/test.dart';
 import 'package:tests/database/database.dart';