+++ extras/integration_tests/tests/lib/suite/transactions.dart
@@ -1,3 +1,4 @@
+import 'package:moor/moor.dart';
 import 'package:test/test.dart';
 import 'package:tests/data/sample_data.dart' as people;
 import 'package:tests/database/database.dart';