+++ moor/lib/src/utils/single_transformer.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 /// Transforms a stream of lists into a stream of single elements, assuming
-/// that each list is a singleton.
-StreamTransformer<List<T>, T?> singleElements<T>() {
+/// that each list is a singleton or empty.
+StreamTransformer<List<T>, T?> singleElementsOrNull<T>() {
   return StreamTransformer.fromHandlers(handleData: (data, sink) {
     try {
       if (data.isEmpty) {
@@ -10,6 +10,19 @@ StreamTransformer<List<T>, T?> singleElements<T>() {
       } else {
         sink.add(data.single);
       }
+    } catch (e) {
+      sink.addError(
+          StateError('Expected exactly one element, but got ${data.length}'));
+    }
+  });
+}
+
+/// Transforms a stream of lists into a stream of single elements, assuming
+/// that each list is a singleton.
+StreamTransformer<List<T>, T> singleElements<T>() {
+  return StreamTransformer.fromHandlers(handleData: (data, sink) {
+    try {
+      sink.add(data.single);
     } catch (e) {
       sink.addError(
           StateError('Expected exactly one element, but got ${data.length}'));