+++ moor/test/ffi/setup_test.dart
@@ -1,5 +1,6 @@
 @TestOn('vm')
 import 'package:moor/ffi.dart';
+import 'package:moor/moor.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
 
@@ -19,7 +20,7 @@ void main() {
     final db = TodoDb(executor);
     final row = await db.customSelect('SELECT my_function() AS r;').getSingle();
 
-    expect(row!.readString('r'), 'hello from Dart');
+    expect(row.readString('r'), 'hello from Dart');
     await db.close();
   });
 
@@ -31,7 +32,7 @@ void main() {
 
     final db = TodoDb(executor);
     final row = await db.customSelect('SELECT my_function() AS r;').getSingle();
-    expect(row!.readString('r'), 'hello from Dart');
+    expect(row.readString('r'), 'hello from Dart');
 
     final nativeRow = existing.select('SELECT my_function() AS r;').single;
     expect(nativeRow['r'], 'hello from Dart');