+++ moor/test/parsed_sql/moor_files_integration_test.dart
@@ -120,7 +120,7 @@ void main() {
   });
 
   test('applies default parameter expressions when not set', () async {
-    await db.readDynamic().getSingle();
+    await db.readDynamic().getSingleOrNull();
 
     verify(mock.runSelect('SELECT * FROM config WHERE (TRUE)', []));
   });